removing duplication from stories and test to make code climate stuff happy
