+++ client/app/nonComp/pages/ClaimHistoryPage.stories.js
@@ -1,32 +1,7 @@
 import React from 'react';
-import CombinedNonCompReducer from '../reducers';
-import { Provider } from 'react-redux';
-import { applyMiddleware, createStore, compose } from 'redux';
-import { MemoryRouter as Router } from 'react-router-dom';
-
-import thunk from 'redux-thunk';
-
 import ClaimHistoryPage from './ClaimHistoryPage';
-
+import ReduxDecorator from 'test/app/nonComp/nonCompReduxDecorator';
 import individualClaimHistoryData from 'test/data/nonComp/individualClaimHistoryData';
-
-const ReduxDecorator = (Story, options) => {
-  const props = {
-    ...options.args.data
-  };
-
-  const store = createStore(
-    CombinedNonCompReducer,
-    props,
-    compose(applyMiddleware(thunk))
-  );
-
-  return <Provider store={store} >
-    <Router>
-      <Story />
-    </Router>
-  </Provider>;
-};
 
 export default {
   title: 'Queue/NonComp/ClaimHistoryPage',
